Converts the world and server stores to be more inline with livewire. Worlds store now dynamically updates based on the world you have selected.
