+++ app/Http/Livewire/Stores/Table.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace App\Http\Livewire\Stores;
+
+use Livewire\Component;
+
+class Table extends Component
+{
+    public $goodRow;
+    public $goodType;
+
+    public function mount()
+    {
+        //stuff
+    }
+
+
+    public function render()
+    {
+        return view('livewire.stores.table');
+    }
+}
