+++ app/Http/Traits/FindingGoods.php
@@ -244,40 +244,6 @@ trait FindingGoods {
 
     }
 
-    /**
-     * @param $goodType
-     * @param $goodId
-     *
-     * @return |null
-     */
-    private function getGoodFromGoodTypeAndGoodId($goodType, $goodId) {
-        switch($goodType->id) {
-            case '2' :
-                $model = Ingots::find($goodId);
-                break;
-            case '1' :
-                $model = Ores::find($goodId);
-                break;
-            case '3' :
-                $model = Components::find($goodId);
-                break;
-            case '4' :
-                $model = Tools::find($goodId);
-                break;
-            case '5' :
-                $model = Ammo::find($goodId);
-                break;
-            case '6' :
-                $model = Bottles::find($goodId);
-                break;
-            default:
-                $model = Tools::find($goodId);
-
-        }
-
-        return (! empty($model)) ? $model : null;
-    }
-
 
     /**
      * @param $goodType
